requirements.txt
Flask
Flask-Cors
Flask-SocketIO
Flask-Migrate
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Mail
python-dotenv
pandas
requests
alembic